refactor(plugin): extract command_recorder.rs + finalize shared tests (#1079)
Moves the RhiCommandRecorderMethodsVTable callbacks +
HOST_RHI_COMMAND_RECORDER_METHODS_VTABLE + accessor +
gpu_rhi_command_recorder_methods_vtable_null_tests out of
host_services/mod.rs into a sibling submodule, completing the
per-vtable extractions.

Also moves the two cross-cutting test modules into their natural
homes:
- run_host_extern_c_panic_safety_net_tests → shared/wire.rs
  (alongside the run_host_extern_c re-export it tests).
- make_borrow_cached_field_regression_tests → shared/borrow.rs
  (alongside the make_*_borrow helpers it tests).

mod.rs is now ~1240 lines — host-services foundation + module tree
+ runtime_facing + RegisterHelper + foundational callbacks only.
Down from 18,687 lines pre-refactor.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
