+++ libs/streamlib-engine/src/core/plugin/host_services/shared/borrow.rs
@@ -260,3 +260,204 @@ pub(in crate::core::plugin::host_services) fn make_graphics_kernel_borrow(
         cached_descriptor_sets_in_flight: inner.descriptor_sets_in_flight(),
     })
 }
+#[cfg(all(test, target_os = "linux"))]
+mod make_borrow_cached_field_regression_tests {
+    //! Locks the issue #988 bug: `make_*_borrow` helpers MUST populate
+    //! the β-shape's cached POD fields from the host-side inner —
+    //! NOT leave them zeroed. Reverting any `make_*_borrow` to
+    //! `width_cached: 0` / `byte_size_cached: 0` / etc. trips these
+    //! assertions.
+    //!
+    //! Requires a working Vulkan device; skips cleanly when one isn't
+    //! available (per `project_ci_strategy_no_gpu`).
+    use super::*;
+    use std::sync::Arc;
+
+    fn try_vulkan_device() -> Option<Arc<crate::vulkan::rhi::HostVulkanDevice>> {
+        crate::vulkan::rhi::HostVulkanDevice::new().ok()
+    }
+
+    #[test]
+    fn make_texture_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        let desc = crate::core::rhi::TextureDescriptor::new(
+            640,
+            480,
+            crate::core::rhi::TextureFormat::Rgba8Unorm,
+        );
+        let host_texture = crate::vulkan::rhi::HostVulkanTexture::new(&device, &desc)
+            .expect("texture allocate");
+        use crate::host_rhi::HostTextureExt;
+        let texture = crate::core::rhi::Texture::from_vulkan(host_texture);
+        let borrow = make_texture_borrow(texture.handle);
+        assert_eq!(borrow.width(), 640, "width_cached must mirror the inner");
+        assert_eq!(borrow.height(), 480, "height_cached must mirror the inner");
+        assert!(
+            matches!(borrow.format(), crate::core::rhi::TextureFormat::Rgba8Unorm),
+            "format_raw must mirror the inner"
+        );
+    }
+
+    #[test]
+    fn make_storage_buffer_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        let host_buffer =
+            crate::vulkan::rhi::HostVulkanBuffer::new_storage_buffer_host_visible(
+                &device, 16_384,
+            )
+            .expect("storage buffer allocate");
+        let buffer = crate::core::rhi::StorageBuffer::from_arc_into_raw(Arc::new(host_buffer));
+        let borrow = make_storage_buffer_borrow(buffer.handle);
+        assert_eq!(borrow.byte_size(), 16_384, "byte_size_cached must mirror the inner");
+        assert!(
+            !borrow.mapped_ptr().is_null(),
+            "mapped_ptr_cached must mirror the inner HOST_VISIBLE pointer"
+        );
+    }
+
+    #[test]
+    fn make_pixel_buffer_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        // Bgra8 = 4 bytes/pixel, 320x240 = 307_200 bytes
+        let host_buffer =
+            crate::vulkan::rhi::HostVulkanBuffer::new_storage_buffer_host_visible(
+                &device, 320 * 240 * 4,
+            )
+            .expect("backing buffer allocate");
+        let pb = crate::core::rhi::PixelBuffer::from_host_vulkan_buffer(
+            Arc::new(host_buffer),
+            320,
+            240,
+            4,
+            crate::core::rhi::PixelFormat::Bgra32,
+        );
+        let borrow = make_pixel_buffer_borrow(pb.handle);
+        assert_eq!(borrow.width, 320, "width must mirror the inner");
+        assert_eq!(borrow.height, 240, "height must mirror the inner");
+        assert!(
+            matches!(borrow.format(), crate::core::rhi::PixelFormat::Bgra32),
+            "format_raw must mirror the inner"
+        );
+    }
+
+    #[test]
+    fn make_uniform_buffer_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        let buffer = crate::core::rhi::UniformBuffer::new_host_visible(&device, 4_096)
+            .expect("uniform buffer allocate");
+        let borrow = make_uniform_buffer_borrow(buffer.handle);
+        assert_eq!(borrow.byte_size(), 4_096, "byte_size_cached must mirror the inner");
+        assert!(
+            !borrow.mapped_ptr().is_null(),
+            "mapped_ptr_cached must mirror the inner HOST_VISIBLE pointer"
+        );
+    }
+
+    #[test]
+    fn make_vertex_buffer_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        let buffer = crate::core::rhi::VertexBuffer::new_host_visible(&device, 8_192)
+            .expect("vertex buffer allocate");
+        let borrow = make_vertex_buffer_borrow(buffer.handle);
+        assert_eq!(borrow.byte_size(), 8_192, "byte_size_cached must mirror the inner");
+        assert!(
+            !borrow.mapped_ptr().is_null(),
+            "mapped_ptr_cached must mirror the inner HOST_VISIBLE pointer"
+        );
+    }
+
+    #[test]
+    fn make_index_buffer_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        let buffer = crate::core::rhi::IndexBuffer::new_host_visible(&device, 2_048)
+            .expect("index buffer allocate");
+        let borrow = make_index_buffer_borrow(buffer.handle);
+        assert_eq!(borrow.byte_size(), 2_048, "byte_size_cached must mirror the inner");
+        assert!(
+            !borrow.mapped_ptr().is_null(),
+            "mapped_ptr_cached must mirror the inner HOST_VISIBLE pointer"
+        );
+    }
+
+    #[test]
+    fn make_compute_kernel_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        // Reuse the test_blend_1 shader already wired in build.rs: one
+        // storage buffer binding at slot 0, one push-constant block of
+        // 4 bytes. The assertion below pins the value the cached field
+        // must mirror.
+        const TEST_BLEND_1_SPV: &[u8] =
+            include_bytes!(concat!(env!("OUT_DIR"), "/test_blend_1.spv"));
+        let descriptor = crate::core::rhi::ComputeKernelDescriptor {
+            label: "make_compute_kernel_borrow_test",
+            spv: TEST_BLEND_1_SPV,
+            bindings: &[
+                crate::core::rhi::ComputeBindingSpec::storage_buffer(0),
+                crate::core::rhi::ComputeBindingSpec::storage_buffer(8),
+            ],
+            push_constant_size: 4,
+        };
+        let kernel =
+            crate::vulkan::rhi::VulkanComputeKernel::new(&device, &descriptor)
+                .expect("compute kernel construct");
+        let borrow = make_compute_kernel_borrow(kernel.handle);
+        assert_eq!(
+            borrow.push_constant_size(),
+            4,
+            "cached_push_constant_size must mirror the inner",
+        );
+    }
+
+    #[test]
+    fn make_acceleration_structure_borrow_populates_cached_pod_fields() {
+        let Some(device) = try_vulkan_device() else {
+            return;
+        };
+        if !device.supports_ray_tracing_pipeline() {
+            return;
+        }
+        // Single triangle BLAS, smallest payload that exercises the
+        // build path. Mirrors the rt-smoke fixture's vertex layout.
+        let vertices: Vec<f32> = vec![
+            0.0, -0.5, 0.0, -0.5, 0.5, 0.0, 0.5, 0.5, 0.0,
+        ];
+        let indices: Vec<u32> = vec![0, 1, 2];
+        let blas = crate::vulkan::rhi::VulkanAccelerationStructure::build_triangles_blas(
+            &device,
+            "make_borrow_test_blas",
+            &vertices,
+            &indices,
+        )
+        .expect("blas construct");
+        let borrow = make_acceleration_structure_borrow(blas.handle);
+        assert!(
+            matches!(
+                borrow.kind(),
+                crate::vulkan::rhi::AccelerationStructureKind::BottomLevel,
+            ),
+            "cached_kind must mirror the inner",
+        );
+        assert!(
+            borrow.device_address() > 0,
+            "cached_device_address must mirror the inner",
+        );
+        assert!(
+            borrow.storage_size() > 0,
+            "cached_storage_size must mirror the inner",
+        );
+    }
+}
